+++ hydra/experimental/compose.py
@@ -1,10 +1,7 @@
 # Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved
 from typing import List, Optional
 
-from omegaconf import DictConfig, open_dict
-
-from hydra.core.global_hydra import GlobalHydra
-from hydra.types import RunMode
+from omegaconf import DictConfig
 
 
 def compose(
@@ -12,31 +9,15 @@ def compose(
     overrides: List[str] = [],
     return_hydra_config: bool = False,
 ) -> DictConfig:
-    """
-    :param config_name: the name of the config
-           (usually the file name without the .yaml extension)
-    :param overrides: list of overrides for config file
-    :param return_hydra_config: True to return the hydra config node in the result
-    :return: the composed config
-    """
-    assert GlobalHydra().is_initialized(), (
-        "GlobalHydra is not initialized, use @hydra.main()"
-        " or call one of the hydra.experimental initialize methods first"
-    )
+    from hydra import compose as real_compose
 
-    gh = GlobalHydra.instance()
-    assert gh.hydra is not None
-    cfg = gh.hydra.compose_config(
+    # warnings.warn(
+    #     category=UserWarning,
+    #     message="hydra.experimental.compose() is no longer experimental."
+    #     " Use hydra.compose()",
+    # )
+    return real_compose(
         config_name=config_name,
         overrides=overrides,
-        run_mode=RunMode.RUN,
-        from_shell=False,
-        with_log_configuration=False,
+        return_hydra_config=return_hydra_config,
     )
-    assert isinstance(cfg, DictConfig)
-
-    if not return_hydra_config:
-        if "hydra" in cfg:
-            with open_dict(cfg):
-                del cfg["hydra"]
-    return cfg
